Reduce max page size and update database configurations
Changed the max page size from 40000 to 10000 in both data rest and pageable sections of spring configurations to enhance system performance. Updated the database configuration to allow dynamic environment variables for flexibility and security. Also added new configuration parameters under the recommendation section of the economatic application for better customization of the application.