Remove unused exception and implement new methods and classes
GoldPriceUpdateRateException.java was deleted as it was not being utilized. Its usage was replaced with an all new exception class, InvalidGoldPricePropertyException.java. Several important enhancements are made to the GoldPriceStatsService class. Additional helper methods are implemented in this service for generating gold price statistics. GoldPriceMapper refactored into an interface from abstract class. To enhance the performance, caching annotations were added to the methods of the GoldPriceService class.