Implemented caching and recommendation service
Implemented CachingConfig for cache management and also a recommendation system. The caching system was refactored to separately cache different data types such as itemPrice, population, goldPrice etc. with respective expiration times. In addition, `EnableCaching` annotation was added to enable Spring's caching capability.

A recommendation system was implemented, which provides recommendations based on item price, population, gold price etc. Exception handling was also added for cases when these properties are negative. The responses are given in sorted order in descending according to their total score. To further refine the recommendations, a limit can be set on the number of recommendations returned and whether to use market value or min buyout for item prices can be chosen.

The changes also include related attributes and methods, as reflected by the newly added .java files. Overall, these changes are crucial for enhancing the performance and user experience of the application.