Refactor gold price related components for simplification
A number of changes were done with the aim of simplifying the gold price related components. The GoldPricesMapper and some related dtos were deleted, as they were deemed unnecessary. Other dtos were renamed for more clarity. In GoldPriceMapper, methods for mapping to GoldPriceListResponse and GoldPricePageResponse were added. Similarly, the GoldPriceStatMapper was changed from an abstract class to an interface to simplify its usage. GoldPriceService was updated to account for these changes.

This refactoring was undertaken to streamline the codebase, provide better clarity on the methods' purposes and reduce redundancy that was previously present.