Remove unused DTOs and Mappers
A significant code clean-up was done, removing unnecessary DTOs, Mappers, and certain service methods no longer in use. Importantly, the methods utilizing these DTOs and mappers have also been updated. The motivation for this change is the simplification and maintaining the clarity of the project architecture. This refactoring makes the project easier to navigate and to continue its development.