Update Json deserialization setting, parallel streaming, Jib version and container name
Modified the Json deserialization settings in `PopulationUpdateService.java` for better handling of unknown properties in incoming JSON. In `ItemPriceUpdateService.java`, replaced 'stream' with 'parallelStream' to improve data processing efficiency for larger datasets. Upgraded Jib version from 3.3.1 to 3.3.2 in `build.gradle` for the latest updates and bug fixes. Changed container name in `docker-compose.yml` to reflect the actual service it represents.