Add additional validation and error handling to application code
The changes in this commit mainly focus on improving the error handling and validation in application code. Validations were added to confirm that specific parameters are non-null before they are used. Also, an explicit exception message was added when no prices are found in the case of empty results. EntityGraph annotations were uncommented to enable eager loading of associated entities, likely improving performance. Also some fields were reordered in a few DTO imports and methods were rearranged to improve readability. The GoldPriceDeserializer was revised to use stream API for creating a list of GoldPriceResponse. Unit tests are still pending.