Extract version management to separate file
Created a separate `version.gradle` file to manage the versions of all dependencies from a central place. This enhances maintainability as it provides an easy way to update all dependencies in a single place rather than scattered around in `build.gradle`. It also minimises the risk of disparate versions across multiple dependencies.