Update Github Actions workflow for Gradle
This commit modifies the Github Actions workflow for building the project with Gradle. The trigger was slightly modified by adding spaces between the brackets; however the key update includes changes for execution permissions of gradlew and use of `gradle/gradle-build-action@v2` to build and test the project. This ensures the gradle wrapper is executable and aligns the workflow with the standard Gradle practices.