Integrate CI/CD in GitHub Actions
Modified the workflow file to rename CI to CI/CD and added a procedure to build and push Docker image. The need for this change arose from the requirement to automate the process of creating Docker images and the push operation to the Docker Hub after every push event on protected branches. This will ensure that there is always the latest, updated Docker image available for deployment.