Refactor test structure and clean up unused code
Deleted redundant test files and SQL scripts that were no longer in use, as they were cluttering the codebase. Repositories and services being tested have been simplified and no longer require complex mocking or database setups in tests.

Renamed the existing test files to clarify that they are unit tests, and created separate directories for unit and integration tests to uphold the standard test structure. This will make it easier to identify and run these different types of tests in the future, improving the overall maintainability of our test suite.

An application-integration.yml file has also been created for setting up application configuration specific to integration tests.

To summarize, all the changes are focused on improving the organization and clarity of tests, allowing easier expansions and maintenance in the future.