Refactor to Specification from SearchSpecification
The class SearchSpecification was refactored and named as SpecificationBuilder, which is implemented as a utility class. This refactoring is done to make the specification creation method stateless and the class itself easier to use across different services. The methods in the class are converted to static methods and the service classes (ServerServiceImpl, GoldPriceService, PopulationService, ItemPriceService, ItemServiceImpl) are updated accordingly.

Also, Unit tests are updated to match this refactoring in GoldPriceServiceTest. Dependencies on the SearchSpecification class in all the services are removed which improves the overall design and readability of the codebase.