Refactor RecommendationService to interface and implemented class
This refactor changes RecommendationService from being a class to an interface, with the original implementation details moved to a new class named BasicRecommendationService. The @Validated annotation was removed from ServerServiceImpl as it was unneeded. The import statements' order was modified for better readability. This update allows for more flexible service integration and will enable easier testing by separating interface and implementation.