Refactor items summary to a new service
Items summary related codes are moved to a new service. Specifically, summary related files including classes, dtos, repositories, and services are removed from the item folder and transferred to a new summary folder. This is done to disentangle the concerns of individual items and summaries. It is a step towards better code organization and readability, and it could improve the maintainability of the codebase in the future.