"Update test profiles and extensions"
Changes have been made to the testing configuration. In ItemDealsRepositoryTest, we removed the "test" profile to exclusively use the "integration" testing profile. This aligns with our new approach of using only integration tests in this context for more realistic testing scenarios. In BaseItemDealTest, we replaced the SpringExtension with MockitoExtension to better fit our test requirements, as we are extensively using Mockito for mocking dependencies in unit testing.