Refactored ItemPriceUpdateService and created NexusHubService
Separated the concerns of ItemPriceUpdateService by creating a new service, NexusHubService. This new service is responsible for communication with the NexusHubClient and related operations, making ItemPriceUpdateService's code cleaner and easier to understand. Removed unnecessary fields and imports from ItemPriceUpdateService and moved methods related to NexusHubClient to NexusHubService. This enhances maintainability and allows for potential unit testing for NexusHubService.