Update code formatting for consistency
This commit includes changes in various files to maintain the consistency in code. Extraneous line breaks have been removed in some files and needed line breaks have been added in others for readability. The sequence of the import statements have been altered in some files for it to be in proper alphabetical order. In files like 'PopulationStatServiceImpl.java', 'GoldPriceStatServiceImpl.java' and more, indentation has been adjusted to maintain consistency across code base. The rearrangement does not change any functionality, it is purely for readability and code maintainability.